requirements.txt
Django==1.3
PIL==1.1.7
South==0.7.3
cmsplugin-blog==1.1.2
cmsplugin-filer==0.8.0
django-admin-tools==0.4.0
django-ajax-selects==1.2.4
django-appmedia==1.0.1
django-classy-tags==0.3.4.1
django-cms==2.2
django-compressor==0.9.2
django-filer==0.8.6
django-ignoretests==0.3.2
django-inline-ordering==0.1.3
django-mailchimp==0.1.30
django-missing==0.1.4
django-model-utils==1.1.0
django-mptt==0.5.1
django-recaptcha-works==0.3.4
django-sekizai==0.5
django-staticfiles==1.0.1
django-tagging==0.3.1
django-tinymce==1.5.1a2
django-xhtml2pdf==0.0.3
djangocms-utils==0.9.5
easy-thumbnails==1.0-alpha-21
elaphe==0.5.3
geojson==1.0.1
google-api-python-client==1.0beta8
html5lib==0.90
httplib2==0.7.4
ipython==0.10.2
oauth2==1.5.211
psycopg2==2.4.1
pyPdf==1.13
pyproj==1.8.9
python-dateutil==2.1
python-gflags==2.0
python-memcached==1.48
readline==6.2.2
reportlab==2.5
simple_translation==0.8.6
simplejson==2.1.6
simplekml==0.7.3
six==1.1.0
-e git://github.com/jdunck/python-unicodecsv.git@22eabfe1cfebed61db8030993d0989ba4ec5a96b#egg=unicodecsv-dev
virtualenv==1.7.1.2
wsgiref==0.1.2
xhtml2pdf==0.0.3
